pyproject.toml
[tool.poetry]
name = "cheers-health"
version = "0.1.0"
description = ""
authors = ["Shivansh Goel <shivanshgoel009@gmail.com>"]
readme = "README.md"
packages = [{include = "cheers_health"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.100.1"}
autopep8 = "^2.0.2"
gradio = "^3.39.0"
openai = "^0.27.8"
pydantic-settings = "^2.0.2"
pandas = "^2.0.3"
openpyxl = "^3.1.2"
sqlalchemy = "^2.0.19"
psycopg2 = "^2.9.7"
uuid = "^1.30"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
